Add user properties and update logout navigation
This commit introduces new properties (`Address`, `City`, `FirstName`, `LastName`, and `Street`) to the `User` model in `User.cs`. The database schema is updated with a new migration (`MIG1`) to add corresponding columns to the `AspNetUsers` table. The `ModelsContextModelSnapshot.cs` file is also updated to reflect these changes. Additionally, the logout functionality in `AccountPage.jsx` is modified to redirect users to a different route upon logout, enhancing user experience.